fix: resolve dtype mismatch, NaN training, HF auth, clone crash
merge.py: delta is float32, weight is bfloat16 -- += raised RuntimeError;
cast delta to weight dtype before accumulation.

Notebook: use Colab Secrets for HF_TOKEN (no hardcoded placeholder);
guard git clone against re-run crash; add grad clipping (NaN risk with
bfloat16 + high LR); show params_before vs after in inject cell.

+++ src/core/merge.py
@@ -15,7 +15,7 @@ def merge_lora_weights(model: nn.Module) -> None:
     for name, module in targets:
         with torch.no_grad():
             delta = (module.lora.A @ module.lora.B) * module.lora.scaling
-            module.linear.weight.data += delta.T
+            module.linear.weight.data += delta.T.to(module.linear.weight.dtype)
             for param in module.linear.parameters():
                 param.requires_grad = True
